requirements.txt
fpdf==1.7.2
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.5.0
Pillow==9.5.0
PyPDF2==3.0.1
Requests==2.31.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
torch==2.0.1
torchvision==0.15.2
torchviz==0.0.2
